终端指数概览分析页面

+++ index-server/src/main/java/com/cmri/um/he/index/terminal/dao/TerminalOverviewDao.java
@@ -46,7 +46,7 @@ public class TerminalOverviewDao extends BaseDao {
      * @param month
      * @return
      */
-    public int quaryTotal(Integer id, String month) {
+    public double quaryTotal(Integer id, String month) {
         return terminalOverviewMapper.quaryTotal(id,month);
 
     }
